backend/project/requirements.txt
django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
requests==2.31.0
psycopg2>=2.9.9
